requirements.txt
fairscale==0.4.13
ftfy==6.3.1
lpips==0.1.4
matplotlib==3.7.2
numpy==2.2.1
omegaconf==2.3.0
opencv_python==4.8.1.78
packaging==24.2
pandas==2.2.3
Pillow==10.0.1
pytest==8.3.4
regex==2024.11.6
Requests==2.32.3
scikit_learn==1.2.2
scipy==1.11.3
setuptools==68.0.0
streamlit==1.39.0
timm==0.9.7
torchsummary==1.5.1
tqdm==4.66.4
tqdm==4.67.0
ultralytics==8.3.56
